refactor: decompose index.ts into modular API route modules
Extract 20 handler functions from src/index.ts (1096 lines) into
focused modules under src/api/ (474 lines remaining, 57% reduction):

- src/api/types.ts — Route, ApiContext, matchPath, serveWorkspaceYaml
- src/api/world-state.ts — world state, services, agent-health, flow, outcomes
- src/api/github.ts — CI health, PR pipeline, branch drift (new)
- src/api/incidents.ts — incident CRUD, security summary
- src/api/operations.ts — ceremonies, skills, channels, HITL, publish, onboard
- src/api/index.ts — barrel export, createAllRoutes()

Each module exports createRoutes(ctx: ApiContext) returning Route[].
To add new endpoints: create a module, import in api/index.ts. No
changes to src/index.ts needed.

Also adds branch drift domain:
- GET /api/branch-drift — commit distance dev→staging→main per project
- Goal: branch.drift_low (max 30 commits)
- Action: alert.branch_drift
- Domain: 10-min tick (GitHub rate limits)
- Quinn tool: get_branch_drift
