+++ src/index.ts
@@ -1,6 +1,6 @@
-import { existsSync, readdirSync, mkdirSync, readFileSync, writeFileSync } from "node:fs";
+import { existsSync, readdirSync, mkdirSync, readFileSync } from "node:fs";
 import { resolve, join, extname } from "node:path";
-import { parse as parseYaml, stringify as stringifyYaml } from "yaml";
+import { parse as parseYaml } from "yaml";
 import { InMemoryEventBus } from "../lib/bus";
 import { DebugPlugin } from "../lib/plugins/debug";
 import { LoggerPlugin } from "../lib/plugins/logger";
@@ -10,18 +10,6 @@ import { SchedulerPlugin } from "../lib/plugins/scheduler";
 import { ActionRegistry } from "./planner/action-registry";
 import type { Plugin, BusMessage } from "../lib/types";
 import type { Action } from "./planner/types/action";
-// Security incident shape — application-level, owned by protoMaker
-interface SecurityIncident {
-  id: string;
-  title: string;
-  severity: "critical" | "high" | "medium" | "low";
-  status: "open" | "investigating" | "resolved";
-  reportedAt: string;
-  description?: string;
-  affectedProjects?: string[];
-  assignee?: string;
-}
-
 // --- Workspace config ---
 const workspaceDir = resolve(
   process.env.WORKSPACE_DIR || join(process.cwd(), "workspace")
@@ -435,575 +423,19 @@ cli?.showPrompt();
 const HTTP_PORT = parseInt(process.env.WORKSTACEAN_HTTP_PORT || "3000", 10);
 const API_KEY = process.env.WORKSTACEAN_API_KEY;
 
-function serveWorkspaceYaml(filename: string, key: string): Response {
-  const filePath = join(workspaceDir, filename);
-  if (!existsSync(filePath)) return Response.json({ success: true, data: [] });
-  try {
-    const parsed = parseYaml(readFileSync(filePath, "utf8")) as Record<string, unknown>;
-    return Response.json({ success: true, data: parsed[key] ?? [] });
-  } catch {
-    return Response.json({ success: false, error: `Failed to parse ${filename}` }, { status: 500 });
-  }
-}
-
-async function handlePublish(req: Request): Promise<Response> {
-  if (API_KEY && req.headers.get("X-API-Key") !== API_KEY) {
-    return Response.json({ success: false, error: "Unauthorized" }, { status: 401 });
-  }
-
-  let body: Record<string, unknown>;
-  try {
-    body = (await req.json()) as Record<string, unknown>;
-  } catch {
-    return Response.json({ success: false, error: "Invalid JSON body" }, { status: 400 });
-  }
-
-  if (!body.topic || typeof body.topic !== "string") {
-    return Response.json(
-      { success: false, error: "Missing or invalid 'topic' (must be a string)" },
-      { status: 400 }
-    );
-  }
-
-  const message: BusMessage = {
-    id: crypto.randomUUID(),
-    correlationId: typeof body.correlationId === "string" ? body.correlationId : crypto.randomUUID(),
-    topic: body.topic,
-    timestamp: Date.now(),
-    payload: body.payload,
-    ...(body.source ? { source: body.source as BusMessage["source"] } : {}),
-    ...(body.reply ? { reply: body.reply as BusMessage["reply"] } : {}),
-  };
-
-  bus.publish(body.topic, message);
-  return Response.json({ success: true, id: message.id });
-}
-
-async function handleOnboard(req: Request): Promise<Response> {
-  if (API_KEY && req.headers.get("X-API-Key") !== API_KEY) {
-    return Response.json({ success: false, error: "Unauthorized" }, { status: 401 });
-  }
-
-  let body: Record<string, unknown>;
-  try {
-    body = (await req.json()) as Record<string, unknown>;
-  } catch {
-    return Response.json({ success: false, error: "Invalid JSON body" }, { status: 400 });
-  }
-
-  if (!body.slug || !body.title || !body.github) {
-    return Response.json(
-      { success: false, error: "Missing required fields: slug, title, github (owner/repo)" },
-      { status: 400 },
-    );
-  }
-
-  const correlationId = crypto.randomUUID();
-  const topic = "message.inbound.onboard";
-
-  // Wait up to 30s for the pipeline to complete and reply
-  const result = await new Promise<Record<string, unknown>>((resolve) => {
-    const replyTopic = `onboard.result.${correlationId}`;
-    let settled = false;
-    let timer: ReturnType<typeof setTimeout>;
-
-    const subId = bus.subscribe(replyTopic, "onboard-http", (msg: BusMessage) => {
-      if (settled) return;
-      settled = true;
-      clearTimeout(timer);
-      bus.unsubscribe(subId);
-      resolve(msg.payload as Record<string, unknown>);
-    });
-
-    timer = setTimeout(() => {
-      if (settled) return;
-      settled = true;
-      bus.unsubscribe(subId);
-      resolve({ success: true, status: "accepted", correlationId, message: "Onboarding started (no response within 30s)" });
-    }, 30_000);
-
-    bus.publish(topic, {
-      id: crypto.randomUUID(),
-      correlationId,
-      topic,
-      timestamp: Date.now(),
-      payload: body,
-      source: { interface: "api" as const },
-      reply: { topic: replyTopic },
-    });
-  });
-
-  const statusCode = result.success ? 200 : 422;
-  return Response.json(result, { status: statusCode });
-}
-
-// ── World-state API helpers ────────────────────────────────────────────────────
-
-// Minimal interface so we can call getWorldState() without importing the full type
-interface WorldStateAPI {
-  getWorldState(opts?: { domain?: string; maxAgeMs?: number }): unknown;
-}
-
-// Resolved after plugins are loaded
-const worldStateCollector = registeredPlugins.find(p => p.name === "world-state-engine") as (WorldStateAPI & { name: string }) | undefined;
-
-function handleGetWorldState(domain?: string): Response {
-  if (!worldStateCollector) {
-    return Response.json({ success: false, error: "world-state-collector not available" }, { status: 503 });
-  }
-  const data = worldStateCollector.getWorldState(domain ? { domain, maxAgeMs: 120_000 } : { maxAgeMs: 120_000 });
-  return Response.json({ success: true, data });
-}
-
-// ── Flow-metrics API ───────────────────────────────────────────────────────────
-
-interface FlowMonitorAPI {
-  getMetrics(opts?: { metric?: string }): unknown;
-}
-
-const flowMonitorPlugin = registeredPlugins.find(p => p.name === "flow-monitor") as (FlowMonitorAPI & { name: string }) | undefined;
-
-function handleGetFlowMetrics(metric?: string): Response {
-  if (!flowMonitorPlugin) {
-    return Response.json({ success: false, error: "flow-monitor not available" }, { status: 503 });
-  }
-  const data = flowMonitorPlugin.getMetrics(metric ? { metric } : undefined);
-  return Response.json({ success: true, data, collectedAt: Date.now() });
-}
-
-// ── Services health API ────────────────────────────────────────────────────────
-
-function handleGetServices(): Response {
-  // Check Discord connectivity via runtime property (private in TS, accessible at runtime)
-  const discord = allPlugins.find(p => p.name === "discord") as Record<string, unknown> | undefined;
-  const discordClient = discord?.["client"] as { isReady(): boolean; user?: { tag: string } } | undefined;
-  const discordReady = discordClient?.isReady() ?? false;
-
-  const services: Record<string, unknown> = {
-    discord: {
-      configured: !!process.env.DISCORD_BOT_TOKEN,
-      connected: discordReady,
-      bot: discordReady ? discordClient?.user?.tag ?? null : null,
-    },
-    github: {
-      configured: !!(process.env.GITHUB_TOKEN || process.env.QUINN_APP_PRIVATE_KEY),
-      authType: process.env.QUINN_APP_PRIVATE_KEY ? "app" : process.env.GITHUB_TOKEN ? "token" : null,
-    },
-    plane: {
-      configured: !!process.env.PLANE_API_KEY,
-      baseUrl: process.env.PLANE_BASE_URL || null,
-    },
-    gateway: {
-      configured: !!process.env.LLM_GATEWAY_URL,
-      url: process.env.LLM_GATEWAY_URL || null,
-    },
-    langfuse: {
-      configured: !!(process.env.LANGFUSE_PUBLIC_KEY && process.env.LANGFUSE_SECRET_KEY),
-    },
-    graphiti: {
-      configured: !!process.env.GRAPHITI_URL,
-      url: process.env.GRAPHITI_URL || null,
-    },
-  };
-
-  return Response.json(services);
-}
-
-// ── Agent health API ──────────────────────────────────────────────────────────
-
-function handleGetAgentHealth(): Response {
-  const registrations = executorRegistry.list();
-
-  // Group by agent name
-  const agents: Record<string, { skills: string[]; executorType: string }> = {};
-  for (const reg of registrations) {
-    const name = reg.agentName ?? "_default";
-    if (!agents[name]) agents[name] = { skills: [], executorType: reg.executor.type };
-    if (reg.skill) agents[name].skills.push(reg.skill);
-  }
-
-  return Response.json({
-    agentCount: Object.keys(agents).length,
-    agents,
-    registrationCount: registrations.length,
-  });
-}
-
-// ── Action outcomes API ───────────────────────────────────────────────────────
-
-interface ActionDispatcherAPI { getOutcomes(): { getAll(): unknown[]; summary(): unknown; getRecent(n: number): unknown[] } }
-
-const actionDispatcher = registeredPlugins.find(p => p.name === "action-dispatcher") as (ActionDispatcherAPI & { name: string }) | undefined;
-
-function handleGetOutcomes(): Response {
-  if (!actionDispatcher) return Response.json({ summary: { success: 0, failure: 0, timeout: 0, total: 0 }, recent: [] });
-  const tracker = actionDispatcher.getOutcomes();
-  return Response.json({
-    summary: tracker.summary(),
-    recent: tracker.getRecent(50),
-  });
-}
-
-// ── CI health + PR pipeline API ────────────────────────────────────────────────
-
-function _loadProjectRepos(): string[] {
-  const projectsPath = join(workspaceDir, "projects.yaml");
-  if (!existsSync(projectsPath)) return [];
-  try {
-    const parsed = parseYaml(readFileSync(projectsPath, "utf8")) as { projects?: Array<{ github?: string }> };
-    return (parsed.projects ?? []).map(p => p.github).filter((g): g is string => !!g);
-  } catch { return []; }
-}
-
-const GITHUB_TOKEN = process.env.GITHUB_TOKEN;
-
-async function _ghApi(path: string): Promise<unknown> {
-  if (!GITHUB_TOKEN) throw new Error("GITHUB_TOKEN not set");
-  const resp = await fetch(`https://api.github.com${path}`, {
-    headers: {
-      Authorization: `Bearer ${GITHUB_TOKEN}`,
-      Accept: "application/vnd.github+json",
-      "X-GitHub-Api-Version": "2022-11-28",
-    },
-    signal: AbortSignal.timeout(15_000),
-  });
-  if (!resp.ok) throw new Error(`GitHub API ${resp.status}: ${path}`);
-  return resp.json();
-}
-
-async function handleGetCiHealth(): Promise<Response> {
-  const repos = _loadProjectRepos();
-  if (!repos.length) return Response.json({ successRate: 1, totalRuns: 0, failedRuns: 0, projects: [] });
-
-  const projects: Array<{ repo: string; successRate: number; totalRuns: number; failedRuns: number; latestConclusion: string | null }> = [];
-  let totalAll = 0;
-  let failedAll = 0;
-
-  for (const repo of repos) {
-    try {
-      const data = await _ghApi(`/repos/${repo}/actions/runs?per_page=10&status=completed`) as {
-        workflow_runs?: Array<{ conclusion: string }>;
-      };
-      const runs = data.workflow_runs ?? [];
-      const failed = runs.filter(r => r.conclusion === "failure").length;
-      const rate = runs.length > 0 ? (runs.length - failed) / runs.length : 1;
-      totalAll += runs.length;
-      failedAll += failed;
-      projects.push({
-        repo,
-        successRate: Math.round(rate * 100) / 100,
-        totalRuns: runs.length,
-        failedRuns: failed,
-        latestConclusion: runs[0]?.conclusion ?? null,
-      });
-    } catch {
-      projects.push({ repo, successRate: 0, totalRuns: 0, failedRuns: 0, latestConclusion: null });
-    }
-  }
-
-  const successRate = totalAll > 0 ? Math.round(((totalAll - failedAll) / totalAll) * 100) / 100 : 1;
-  return Response.json({ successRate, totalRuns: totalAll, failedRuns: failedAll, projects });
-}
-
-async function handleGetPrPipeline(): Promise<Response> {
-  const repos = _loadProjectRepos();
-  if (!repos.length) return Response.json({ totalOpen: 0, conflicting: 0, stale: 0, failing: 0, prs: [] });
-
-  const allPrs: Array<{
-    repo: string; number: number; title: string; mergeable: string;
-    checksPass: boolean; updatedAt: string; stale: boolean;
-  }> = [];
-
-  const sevenDaysAgo = Date.now() - 7 * 24 * 60 * 60 * 1000;
-
-  for (const repo of repos) {
-    try {
-      const data = await _ghApi(`/repos/${repo}/pulls?state=open&per_page=30&sort=updated&direction=desc`) as Array<{
-        number: number; title: string; mergeable_state: string; updated_at: string;
-      }>;
-      for (const pr of data) {
-        const updatedMs = new Date(pr.updated_at).getTime();
-        const isStale = updatedMs < sevenDaysAgo;
-        // Check CI status via commit status
-        let checksPass = true;
-        try {
-          const checks = await _ghApi(`/repos/${repo}/pulls/${pr.number}/reviews`) as Array<{ state: string }>;
-          // Simplified: if any review requests changes, mark as not passing
-          checksPass = !checks.some(r => r.state === "CHANGES_REQUESTED");
-        } catch { /* treat as passing if we can't check */ }
-
-        allPrs.push({
-          repo, number: pr.number, title: pr.title,
-          mergeable: pr.mergeable_state,
-          checksPass, updatedAt: pr.updated_at, stale: isStale,
-        });
-      }
-    } catch { /* skip repos we can't reach */ }
-  }
-
-  const conflicting = allPrs.filter(p => p.mergeable === "dirty").length;
-  const stale = allPrs.filter(p => p.stale).length;
-  const failing = allPrs.filter(p => !p.checksPass).length;
-
-  return Response.json({
-    totalOpen: allPrs.length,
-    conflicting,
-    stale,
-    failing,
-    prs: allPrs,
-  });
-}
-
-function handleGetCeremonies(): Response {
-  const ceremoniesDir = join(workspaceDir, "ceremonies");
-  if (!existsSync(ceremoniesDir)) return Response.json({ success: true, data: [] });
-  const files = readdirSync(ceremoniesDir).filter(f => f.endsWith(".yaml") || f.endsWith(".yml"));
-  const ceremonies: unknown[] = [];
-  for (const file of files) {
-    try {
-      const parsed = parseYaml(readFileSync(join(ceremoniesDir, file), "utf8")) as Record<string, unknown>;
-      ceremonies.push(parsed);
-    } catch { /* skip malformed files */ }
-  }
-  return Response.json({ success: true, data: ceremonies });
-}
-
-async function handleRunCeremony(req: Request, ceremonyId: string): Promise<Response> {
-  if (API_KEY && req.headers.get("X-API-Key") !== API_KEY) {
-    return Response.json({ success: false, error: "Unauthorized" }, { status: 401 });
-  }
-  // Sanitize ceremonyId — only allow alphanumeric, dash, dot, underscore
-  if (!/^[\w.\-]+$/.test(ceremonyId)) {
-    return Response.json({ success: false, error: "Invalid ceremony id" }, { status: 400 });
-  }
-  const topic = `ceremony.${ceremonyId}.execute`;
-  bus.publish(topic, {
-    id: crypto.randomUUID(),
-    correlationId: crypto.randomUUID(),
-    topic,
-    timestamp: Date.now(),
-    payload: { type: "manual.execute", triggeredBy: "api" },
-  });
-  return Response.json({ success: true, message: `Ceremony "${ceremonyId}" triggered` });
-}
-
-// ── Incidents API ─────────────────────────────────────────────────────────────
-
-function handleGetIncidents(): Response {
-  return serveWorkspaceYaml("incidents.yaml", "incidents");
-}
-
-function handleGetSecuritySummary(): Response {
-  const filePath = join(workspaceDir, "incidents.yaml");
-  if (!existsSync(filePath)) return Response.json({ openCount: 0, criticalCount: 0, incidents: [] });
-  try {
-    const parsed = parseYaml(readFileSync(filePath, "utf8")) as { incidents?: SecurityIncident[] };
-    const incidents = parsed.incidents ?? [];
-    const open = incidents.filter(i => i.status !== "resolved");
-    const critical = open.filter(i => i.severity === "critical");
-    return Response.json({
-      openCount: open.length,
-      criticalCount: critical.length,
-      incidents: open.map(i => ({ id: i.id, title: i.title, severity: i.severity, status: i.status })),
-    });
-  } catch {
-    return Response.json({ openCount: 0, criticalCount: 0, incidents: [] });
-  }
-}
-
-async function handleReportIncident(req: Request): Promise<Response> {
-  if (API_KEY && req.headers.get("X-API-Key") !== API_KEY) {
-    return Response.json({ success: false, error: "Unauthorized" }, { status: 401 });
-  }
-
-  let body: Record<string, unknown>;
-  try {
-    body = (await req.json()) as Record<string, unknown>;
-  } catch {
-    return Response.json({ success: false, error: "Invalid JSON body" }, { status: 400 });
-  }
-
-  if (!body.title || !body.severity) {
-    return Response.json(
-      { success: false, error: "Missing required fields: title, severity" },
-      { status: 400 }
-    );
-  }
-
-  const incidentsPath = join(workspaceDir, "incidents.yaml");
-  let existing: SecurityIncident[] = [];
-  if (existsSync(incidentsPath)) {
-    try {
-      const parsed = parseYaml(readFileSync(incidentsPath, "utf8")) as { incidents?: SecurityIncident[] };
-      existing = parsed.incidents ?? [];
-    } catch { /* start fresh */ }
-  }
-
-  const incident: SecurityIncident = {
-    id: `INC-${String(existing.length + 1).padStart(3, "0")}`,
-    title: body.title as string,
-    severity: body.severity as SecurityIncident["severity"],
-    status: (body.status as SecurityIncident["status"]) ?? "open",
-    reportedAt: new Date().toISOString(),
-    ...(body.description ? { description: body.description as string } : {}),
-    ...(Array.isArray(body.affectedProjects) ? { affectedProjects: body.affectedProjects as string[] } : {}),
-    ...(body.assignee ? { assignee: body.assignee as string } : {}),
-  };
-
-  existing.push(incident);
-  writeFileSync(incidentsPath, stringifyYaml({ incidents: existing }), "utf8");
-
-  // Notify bus — world-state-collector re-collects security domain immediately
-  bus.publish("security.incident.reported", {
-    id: crypto.randomUUID(),
-    correlationId: crypto.randomUUID(),
-    topic: "security.incident.reported",
-    timestamp: Date.now(),
-    payload: { incident },
-  });
-
-  return Response.json({ success: true, data: incident }, { status: 201 });
-}
-
-async function handleResolveIncident(req: Request, incidentId: string): Promise<Response> {
-  if (API_KEY && req.headers.get("X-API-Key") !== API_KEY) {
-    return Response.json({ success: false, error: "Unauthorized" }, { status: 401 });
-  }
-
-  const incidentsPath = join(workspaceDir, "incidents.yaml");
-  if (!existsSync(incidentsPath)) {
-    return Response.json({ success: false, error: "No incidents file found" }, { status: 404 });
-  }
-
-  let incidents: SecurityIncident[];
-  try {
-    const parsed = parseYaml(readFileSync(incidentsPath, "utf8")) as { incidents?: SecurityIncident[] };
-    incidents = parsed.incidents ?? [];
-  } catch {
-    return Response.json({ success: false, error: "Failed to parse incidents.yaml" }, { status: 500 });
-  }
-
-  const idx = incidents.findIndex(i => i.id === incidentId);
-  if (idx === -1) {
-    return Response.json({ success: false, error: `Incident "${incidentId}" not found` }, { status: 404 });
-  }
-
-  incidents[idx] = { ...incidents[idx], status: "resolved" };
-  writeFileSync(incidentsPath, stringifyYaml({ incidents }), "utf8");
-
-  bus.publish("security.incident.reported", {
-    id: crypto.randomUUID(),
-    correlationId: crypto.randomUUID(),
-    topic: "security.incident.reported",
-    timestamp: Date.now(),
-    payload: { incident: incidents[idx] },
-  });
-
-  return Response.json({ success: true, data: incidents[idx] });
-}
-
-// ── Channels API ──────────────────────────────────────────────────────────────
-
-function handleGetChannels(): Response {
-  return Response.json({ success: true, data: channelRegistry.getAll() });
-}
-
-async function handleAddChannel(req: Request): Promise<Response> {
-  if (API_KEY && req.headers.get("X-API-Key") !== API_KEY) {
-    return Response.json({ success: false, error: "Unauthorized" }, { status: 401 });
-  }
-
-  let body: Record<string, unknown>;
-  try {
-    body = (await req.json()) as Record<string, unknown>;
-  } catch {
-    return Response.json({ success: false, error: "Invalid JSON" }, { status: 400 });
-  }
-
-  if (!body.id || !body.platform) {
-    return Response.json({ success: false, error: "id and platform are required" }, { status: 400 });
-  }
-
-  try {
-    channelRegistry.add(body as unknown as import("../lib/types/channels.js").Channel);
-    return Response.json({ success: true, data: body });
-  } catch (err) {
-    const msg = err instanceof Error ? err.message : String(err);
-    return Response.json({ success: false, error: msg }, { status: 409 });
-  }
-}
-
-function handleGetHITLPending(): Response {
-  return Response.json({ success: true, data: hitlPlugin.getPendingRequests() });
-}
-
-function handleGetAgentSkills(agentName: string): Response {
-  const agentsPath = join(workspaceDir, "agents.yaml");
-  if (!existsSync(agentsPath)) {
-    return Response.json({ success: false, error: "agents.yaml not found" }, { status: 404 });
-  }
-  try {
-    const parsed = parseYaml(readFileSync(agentsPath, "utf8")) as {
-      agents?: Array<{ name: string; skills?: unknown[] }>;
-    };
-    const agent = (parsed.agents ?? []).find(a => a.name === agentName);
-    if (!agent) {
-      return Response.json({ success: false, error: `Agent "${agentName}" not found` }, { status: 404 });
-    }
-    return Response.json({ success: true, data: { name: agent.name, skills: agent.skills ?? [] } });
-  } catch {
-    return Response.json({ success: false, error: "Failed to parse agents.yaml" }, { status: 500 });
-  }
-}
-
-// ── HTTP router ───────────────────────────────────────────────────────────────
-
-type Params = Record<string, string>;
-type RouteHandler = (req: Request, params: Params) => Response | Promise<Response>;
-
-/** Match a path against a pattern like "/api/foo/:id/run". Returns params or null. */
-function matchPath(pattern: string, path: string): Params | null {
-  const pp = pattern.split("/");
-  const sp = path.split("/");
-  if (pp.length !== sp.length) return null;
-  const params: Params = {};
-  for (let i = 0; i < pp.length; i++) {
-    if (pp[i].startsWith(":")) {
-      params[pp[i].slice(1)] = sp[i];
-    } else if (pp[i] !== sp[i]) {
-      return null;
-    }
-  }
-  return params;
-}
-
-const routes: Array<{ method: string; path: string; handler: RouteHandler }> = [
-  { method: "GET",  path: "/health",                    handler: () => Response.json({ status: "ok", timestamp: Date.now() }) },
-  { method: "POST", path: "/publish",                   handler: handlePublish },
-  { method: "POST", path: "/api/onboard",               handler: handleOnboard },
-  { method: "GET",  path: "/api/projects",              handler: () => serveWorkspaceYaml("projects.yaml", "projects") },
-  { method: "GET",  path: "/api/agents",                handler: () => serveWorkspaceYaml("agents.yaml", "agents") },
-  { method: "GET",  path: "/api/goals",                 handler: () => serveWorkspaceYaml("goals.yaml", "goals") },
-  { method: "GET",  path: "/api/world-state",           handler: () => handleGetWorldState() },
-  { method: "GET",  path: "/api/world-state/:domain",   handler: (_, p) => handleGetWorldState(p.domain) },
-  { method: "GET",  path: "/api/services",              handler: () => handleGetServices() },
-  { method: "GET",  path: "/api/agent-health",          handler: () => handleGetAgentHealth() },
-  { method: "GET",  path: "/api/flow-metrics",          handler: () => handleGetFlowMetrics() },
-  { method: "GET",  path: "/api/flow-metrics/:metric",  handler: (_, p) => handleGetFlowMetrics(p.metric) },
-  { method: "GET",  path: "/api/outcomes",              handler: () => handleGetOutcomes() },
-  { method: "GET",  path: "/api/ci-health",             handler: () => handleGetCiHealth() },
-  { method: "GET",  path: "/api/pr-pipeline",           handler: () => handleGetPrPipeline() },
-  { method: "GET",  path: "/api/ceremonies",            handler: () => handleGetCeremonies() },
-  { method: "POST", path: "/api/ceremonies/:id/run",    handler: (req, p) => handleRunCeremony(req, p.id) },
-  { method: "GET",  path: "/api/skills/:agentName",     handler: (_, p) => handleGetAgentSkills(p.agentName) },
-  { method: "GET",  path: "/api/incidents",            handler: () => handleGetIncidents() },
-  { method: "GET",  path: "/api/security-summary",    handler: () => handleGetSecuritySummary() },
-  { method: "POST", path: "/api/incidents",            handler: (req) => handleReportIncident(req) },
-  { method: "POST", path: "/api/incidents/:id/resolve", handler: (req, p) => handleResolveIncident(req, p.id) },
-  { method: "GET",  path: "/api/channels",             handler: () => handleGetChannels() },
-  { method: "POST", path: "/api/channels",             handler: (req) => handleAddChannel(req) },
-  { method: "GET",  path: "/api/hitl/pending",         handler: () => handleGetHITLPending() },
-];
+// ── API routes (modular) ──────────────────────────────────────────────────────
+import { createAllRoutes, matchPath } from "./api/index.ts";
+import type { ApiContext } from "./api/index.ts";
+
+const apiContext: ApiContext = {
+  workspaceDir,
+  bus,
+  plugins: allPlugins,
+  executorRegistry,
+  apiKey: API_KEY,
+};
+
+const routes = createAllRoutes(apiContext);
 
 Bun.serve({
   port: HTTP_PORT,
@@ -1035,7 +467,8 @@ console.log(`HTTP API listening on port ${HTTP_PORT}`);
     engine.registerDomain("security", createHttpCollector(`${base}/api/security-summary`), 60_000);
     engine.registerDomain("ci", createHttpCollector(`${base}/api/ci-health`), 300_000); // 5min — GitHub rate limits
     engine.registerDomain("pr_pipeline", createHttpCollector(`${base}/api/pr-pipeline`), 120_000); // 2min
+    engine.registerDomain("branch_drift", createHttpCollector(`${base}/api/branch-drift`), 600_000); // 10min
 
-    console.log("[domain-discovery] Registered local domains: flow, services, agent_health, security, ci, pr_pipeline");
+    console.log("[domain-discovery] Registered local domains: flow, services, agent_health, security, ci, pr_pipeline, branch_drift");
   }
 }
