+++ workspace/agents/quinn.yaml
@@ -29,6 +29,7 @@ tools:
   - get_projects
   - get_ci_health
   - get_pr_pipeline
+  - get_branch_drift
   - get_incidents
   - report_incident
   - get_outcomes
