+++ workspace/goals.yaml
@@ -66,6 +66,13 @@ goals:
     max: 0
     description: "No PRs should be stale (>7 days without activity)"
 
+  - id: branch.drift_low
+    type: Threshold
+    severity: medium
+    selector: "domains.branch_drift.data.maxDrift"
+    max: 30
+    description: "No project should have more than 30 commits of branch drift"
+
   # ── Ava / protoMaker Studio ──────────────────────────────────────
   - id: ava.board_healthy
     type: Threshold
